refactor packaging for interface adapters
