--- src/interface_adapter/ListingState.java
@@ -1,6 +1,0 @@
-package interface_adapter;
-
-public class ListingState {
-    // do we add the same input data + property details
-
-}
