+++ src/main/app/HomeSearchUseCaseFactory.java
@@ -4,9 +4,9 @@ import data_access.HomeSearchDataAccessInterface;
 import data_access.InMemoryDataAccessObject;
 import data_access.PropertyDataAccessObject;
 import entity.PropertyFactory;
-import interface_adapter.HomeSearchController;
-import interface_adapter.HomeSearchPresenter;
-import interface_adapter.HomeSearchViewModel;
+import interface_adapter.homeSearch.HomeSearchController;
+import interface_adapter.homeSearch.HomeSearchPresenter;
+import interface_adapter.homeSearch.HomeSearchViewModel;
 import interface_adapter.ViewManagerModel;
 import use_case.home.*;
 import view.HomeSearchView;
